==> notebooks/00_exploration.ipynb <==
DATA_PATH = 'C:/Users/M/Documents/data_science_project/data/raw/survey_results_public.csv'

import pandas as pd
import numpy as np

raw_df = pd.read_csv(DATA_PATH)
raw_df.head()

# Display random answer 
# Observations: Multiple answers need to be splitted 
# Reference to the schema needed to understand 
raw_df.sample(1).iloc[0]

raw_df.info()

raw_df.describe()

questionable_columns = ['YearsCodePro' ,'YearsCode' ,'Age1stCode']
for col in questionable_columns :
    print(col)
    print(raw_df[col].unique().tolist())
    print('----------------')
