==> notebooks/05a_prediction_notebook.ipynb <==
LOG_DATA_PKL = 'data.pkl'
LOG_MODEL_PKL = 'model.pkl'
artifacts_path = 'C:/Users/M/Documents/data_science_project/notebooks/mlruns/1/67403f4417c34d82b102947d9e1aaa9d/artifacts/'

import os 
import sklearn
import pickle
import yaml

import pandas as pd

import mlflow
from mlflow.tracking import MlflowClient

# Load Model

# Load data pkl
data_path  = os.path.join(artifacts_path ,LOG_DATA_PKL)
with open(data_path, 'rb') as handle:
    data_pkl = pickle.load(handle)

# Load model pkl
model_path = os.path.join(artifacts_path, LOG_MODEL_PKL)
with open(model_path, 'rb') as handle:
    model_pkl = pickle.load(handle)

model = model_pkl["model_object"]

model

# Predict Sample Entry 

CLUSTERS_YAML_PATH = "C:/Users/M/Documents/data_science_project/data/processed/features_skills_clusters_description.yaml"

CLUSTERS_YAML_PATH

with open(CLUSTERS_YAML_PATH ,'r') as stream :
    cluster_config = yaml.safe_load(stream)

cluster_config

molten_clusters = [(cluster_name, cluster_skill)
                   for cluster_name, cluster_skills in cluster_config.items()
                   for cluster_skill in cluster_skills]

clusters_df = pd.DataFrame(molten_clusters, columns=["cluster_name", "skill"])
clusters_df

# Recreate cluster features

sample_skills = ['Pandas', 'TensorFlow', 'Torch/PyTorch', 'Python', 'Keras']

sample_clusters = clusters_df.copy()

sample_clusters["sample_skills"] = sample_clusters["skill"].isin(sample_skills)

cluster_features = sample_clusters.groupby('cluster_name')['sample_skills'].sum()
cluster_features

# Create OneHotEncoded Skills

features_names = pd.Series(data_pkl['features_names'])
features_names

skills_names = features_names[~features_names.isin(cluster_features.index)]
skills_names

ohe_skills = pd.Series(skills_names.isin(sample_skills).astype(int).tolist() ,index = skills_names)
ohe_skills

features = pd.concat([ohe_skills ,cluster_features])

features

# Predict

predictions = model.predict_proba([features.values])

positive_pred = [prob[0][1] for prob in predictions]
preds = pd.Series(positive_pred ,index = data_pkl['targets_names']).sort_values(ascending = False)
preds
